lib: Change data generation
For the iid cases, only the function signature changed
For the dependent cases, we now generate a train, a validation and a test set
such that the three sets are independent and they start at the same (random)
initial position.
This way, they should be as similar as possible while maintaining the dependence
structure.